Refactor: 스웨거 셋업 분리
