+++ src/main.ts
@@ -1,20 +1,15 @@
 import { NestFactory } from '@nestjs/core';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { AppModule } from './app.module';
+import { setSwagger } from './setSwagger';
 
 declare const module: any;
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  await app.listen(3000);
 
-  const config = new DocumentBuilder()
-    .setTitle('PayHere Task')
-    .setDescription('by 3rd-wanted-pre-onboarding-team-D')
-    .setVersion('1.0')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document);
+  setSwagger(app);
+
+  await app.listen(3000);
 
   if (module.hot) {
     module.hot.accept();
